# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Blytz.live Marketplace API
  description: |
    Comprehensive REST API for Blytz.live Marketplace - Live Auction Platform
    
    Features:
    - User Authentication & Management
    - Product Catalog with Categories
    - Shopping Cart & Order Management
    - Live Auction System
    - Real-time Bidding
    - Admin Dashboard
    
    All times are in UTC unless otherwise specified.
  version: 3.0.0
  contact:
    name: Blytz Ventures Support
    email: support@blytzventures.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.blytz.live/api/v1
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /products/{id}:
    get:
      tags:
        - Products
      summary: Get product by ID
      description: Retrieve product details by ID
      operationId: getProductByID
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Products
      summary: Update product
      description: Update existing product (owner/admin only)
      operationId: updateProduct
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdateRequest'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not product owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Products
      summary: Delete product
      description: Soft delete product (owner/admin only)
      operationId: deleteProduct
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '403':
          description: Forbidden - not product owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
  schemas:
...
    ProductUpdateRequest:
      type: object
      properties:
        category_id:
          type: string
          format: uuid
          description: Category ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        title:
          type: string
          description: Product title
          example: "iPhone 15 Pro"
          minLength: 3
          maxLength: 200
        description:
          type: string
          nullable: true
          description: Product description
          example: "Latest iPhone with advanced features"
        condition:
          type: string
          enum: [new, like_new, good, fair]
          nullable: true
          description: Product condition
          example: "new"
        starting_price:
          type: number
          format: float
          description: Starting price for auction
          example: 999.99
          minimum: 0
        reserve_price:
          type: number
          format: float
          nullable: true
          description: Reserve price (must be >= starting_price)
          example: 1200.00
          minimum: 0
        buy_now_price:
          type: number
          format: float
          nullable: true
          description: Buy now price (must be > starting_price)
          example: 1500.00
          minimum: 0
        images:
          type: array
          items:
            type: string
            format: uri
          description: Array of image URLs
          example: ["https://example.com/image1.jpg", "https://example.com/image2.jpg"]
        video_url:
          type: string
          format: uri
          nullable: true
          description: Product video URL
          example: "https://example.com/video.mp4"
        specifications:
          type: object
          nullable: true
          description: Product specifications as JSON object
          example: {"brand": "Apple", "model": "iPhone 15 Pro", "storage": "256GB"}
        shipping_info:
          type: object
          nullable: true
          description: Shipping information as JSON object
          example: {"weight": 0.5, "dimensions": {"length": 15, "width": 8, "height": 1}}
        status:
          type: string
          enum: [draft, active, sold, cancelled]
          description: Product status
          example: "active"
        featured:
          type: boolean
          description: Featured product flag
          example: false
    ProductResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Product ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        seller_id:
          type: string
          format: uuid
          description: Seller ID
          example: "550e8400-e29b-41d4-a716-446655440001"
        category_id:
          type: string
          format: uuid
          description: Category ID
          example: "550e8400-e29b-41d4-a716-446655440002"
        title:
          type: string
          description: Product title
          example: "iPhone 15 Pro"
        description:
          type: string
          nullable: true
          description: Product description
          example: "Latest iPhone with advanced features"
        condition:
          type: string
          enum: [new, like_new, good, fair]
          nullable: true
          description: Product condition
          example: "new"
        starting_price:
          type: number
          format: float
          description: Starting price for auction
          example: 999.99
        reserve_price:
          type: number
          format: float
          nullable: true
          description: Reserve price
          example: 1200.00
        buy_now_price:
          type: number
          format: float
          nullable: true
          description: Buy now price
          example: 1500.00
        images:
          type: array
          items:
            type: string
            format: uri
          description: Array of image URLs
          example: ["https://example.com/image1.jpg"]
        video_url:
          type: string
          format: uri
          nullable: true
          description: Product video URL
          example: "https://example.com/video.mp4"
        specifications:
          type: object
          nullable: true
          description: Product specifications
          example: {"brand": "Apple", "model": "iPhone 15 Pro"}
        shipping_info:
          type: object
          nullable: true
          description: Shipping information
          example: {"weight": 0.5, "dimensions": {"length": 15, "width": 8, "height": 1}}
        status:
          type: string
          enum: [draft, active, sold, cancelled]
          description: Product status
          example: "active"
        featured:
          type: boolean
          description: Featured product flag
          example: false
        view_count:
          type: integer
          description: Number of views
          example: 42
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2024-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2024-01-01T00:00:00Z"
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid input"
        details:
          type: string
          description: Detailed error information
          example: "Email field is required"
...